requirements.txt
anyio==3.6.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
orjson==3.8.6
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
praw==7.6.1
prawcore==2.3.0
pydantic==1.10.5
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
update-checker==0.18.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==1.5.1
websockets==10.4
